requirements.txt
# requirements.txt

# Tkinter ist in den meisten Python-Installationen enthalten,
# aber unter Linux ggf. separat über das System installiert werden.
# Dieser Platzhalter stellt sicher, dass es dokumentiert ist.
tk

# Requires Python >= 3.10


#Übung 4
numpy~=1.26.4
matplotlib~=3.9.0

#Übung 4.1

numpy
pillow
matplotlib